Simplify template rendering
